# inline get_pixel and get_pixel_mut for >2x speedup