@qarmin 2022-07-23
How much bigger is binary with this code?