@drewcassidy 2022-07-23
The `scaledown` example binary on aarch64 goes from 3,133,143 bytes to 3,150,983 bytes, only a 18kB difference