@fintelia 2022-07-23
Thanks for suggesting this optimization!