@qarmin 2022-07-24
I tested this on Ubuntu 22.04 and i7 4770 and improvements are visible with release mode(maybe not much as in M1, but still big)
Maybe there 

this PR
```
Scaled by near in 21 ms
Scaled by tri in 90 ms
Scaled by cmr in 161 ms
Scaled by gauss in 239 ms
Scaled by lcz2 in 241 ms
Thumbnailed to 20 in 20 ms
Thumbnailed to 40 in 20 ms
Thumbnailed to 100 in 20 ms
Thumbnailed to 200 in 18 ms
Thumbnailed to 400 in 20 ms
```

old master
```
Scaled by near in 24 ms
Scaled by tri in 120 ms
Scaled by cmr in 229 ms
Scaled by gauss in 333 ms
Scaled by lcz2 in 340 ms
Thumbnailed to 20 in 35 ms
Thumbnailed to 40 in 36 ms
Thumbnailed to 100 in 35 ms
Thumbnailed to 200 in 36 ms
Thumbnailed to 400 in 38 ms
```

I also added `#[inline]` into almost every function in this file to test if there is still room for improvements, but I got only performance increased by ~5%  - scaledown example binary increased from 7,238 MB to 7,242 MB
```
Scaled by near in 20 ms
Scaled by tri in 88 ms
Scaled by cmr in 159 ms
Scaled by gauss in 235 ms
Scaled by lcz2 in 235 ms
Thumbnailed to 20 in 18 ms
Thumbnailed to 40 in 18 ms
Thumbnailed to 100 in 19 ms
Thumbnailed to 200 in 17 ms
Thumbnailed to 400 in 19 ms
```